reconcile out of sync files on send-only if only version is different
